Coding under pressure.
This is the raw, ugly, unsafe code I wrote to get the spec to work as
fast as possible. The code before refactoring and cleaning up. With
poor error handling to boot.
Curious to run it with `-race` detector and see if there any race
conditions (I tried to avoid those).
First time using html/template (I've heard of them, but never got a
chance to use), so that was neat to practice.
It was the first time I wrote `struct{}{}` to initialize an empty
struct. But it's very consistent to initializing empty slices, etc.
Just type{}, in this case the type is struct{} instead of []int.
Props to @kluyg for collaboration and getting the peer link sharing to
work together with me!
Sorry for dot importing a gist. :(